+++ addons/tutor_management/models/student.py
@@ -1,4 +1,4 @@
-from odoo import api, fields, models
+from odoo import fields, models
 
 
 class HospitalPatient(models.Model):
@@ -9,7 +9,7 @@ class HospitalPatient(models.Model):
     lesson_ids = fields.One2many(
         comodel_name='tutor.lesson',
         inverse_name='student_id',
-        string='Lesson history',
+        string='Lesson History',
         readonly=True,
     )
     lesson_count = fields.Integer(compute='_compute_lesson_count')
@@ -17,7 +17,3 @@ class HospitalPatient(models.Model):
     def _compute_lesson_count(self):
         for student in self:
             student.lesson_count = len(student.lesson_ids)
-
-    @api.model
-    def action_empty(self, *args, **kwargs):
-        return False